--- src/main/java/ch/heigvd/res/chill/domain/wasadigi/Boxer.java
@@ -1,21 +1,0 @@
-package ch.heigvd.res.chill.domain.wasadigi;
-
-import ch.heigvd.res.chill.domain.IProduct;
-
-import java.math.BigDecimal;
-
-public class Boxer implements IProduct {
-
-  public final static String NAME = "Boxer";
-  public final static BigDecimal PRICE = new BigDecimal(2.0);
-
-  @Override
-  public String getName() {
-    return NAME;
-  }
-
-  @Override
-  public BigDecimal getPrice() {
-    return PRICE;
-  }
-}
