--- src/test/java/ch/heigvd/res/chill/domain/wasadigi/DuvelTest.java
@@ -1,31 +1,0 @@
-package ch.heigvd.res.chill.domain.wasadigi;
-
-import ch.heigvd.res.chill.domain.Bartender;
-import ch.heigvd.res.chill.protocol.OrderRequest;
-import ch.heigvd.res.chill.protocol.OrderResponse;
-import org.junit.jupiter.api.Test;
-
-import java.math.BigDecimal;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-class DuvelTest {
-
-  @Test
-  void thePriceAndNameForPunkIPAShouldBeCorrect() {
-    Duvel beer = new Duvel();
-    assertEquals(beer.getName(), Duvel.NAME);
-    assertEquals(beer.getPrice(), Duvel.PRICE);
-  }
-
-  @Test
-  void aBartenderShouldAcceptAnOrderForPunkIPA() {
-    Bartender jane = new Bartender();
-    String productName = "ch.heigvd.res.chill.domain.wasadigi.Duvel";
-    OrderRequest request = new OrderRequest(3, productName);
-    OrderResponse response = jane.order(request);
-    BigDecimal expectedTotalPrice = Duvel.PRICE.multiply(new BigDecimal(3));
-    assertEquals(expectedTotalPrice, response.getTotalPrice());
-  }
-
-}
